requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
channels==2.4.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colorlover==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==2.5.0
dash==2.6.2
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
Django==4.1.2
django-adaptors==0.2.5
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-perms==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dpd-components==0.1.0
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.13
hyperlink==21.0.0
identify==2.5.6
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandas-datareader==0.10.0
platformdirs==2.5.2
plotly==5.11.0
pre-commit==2.20.0
prompt-toolkit==3.0.31
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
zope.interface==5.5.0
